feat: added email notification for [invites, registering, events remaining]

+++ src/invites/invites.service.ts
@@ -1,15 +1,20 @@
 import { Injectable } from '@nestjs/common';
 import { PrismaService } from 'src/prisma.service';
 import { UserService } from 'src/user/user.service';
+import { CreateInviteDto } from './dto/create-invite.dto';
+import { InjectQueue } from '@nestjs/bullmq';
+import { Queue } from 'bullmq';
 
 @Injectable()
 export class InvitesService {
   constructor(
     private prisma: PrismaService,
-    private userService: UserService
+    private userService: UserService,
+    @InjectQueue('emailQueue') private emailQueue: Queue
   ) {}
 
-  async createInvite(calendarId: string, email: string) {
+  async createInvite(dto: CreateInviteDto) {
+    const { email, calendarId } = dto;
     const user = await this.userService.findByEmail(email);
     if (!user) {
       throw new Error('User not found');
@@ -26,13 +31,26 @@ export class InvitesService {
       return existing;
     }
 
-    return this.prisma.invite.create({
+    const invite = await this.prisma.invite.create({
       data: {
         calendarId,
         email,
         userId: user.id
       }
     });
+
+    const link = `${process.env.FRONTEND_URL}/invites/`;
+
+    await this.emailQueue.add('sendEmail', {
+      to: email,
+      subject: 'You have been invited to a calendar',
+      template: 'invite',
+      context: {
+        link
+      }
+    });
+
+    return invite;
   }
 
   async acceptInvite(id: string, userId: string) {
@@ -70,7 +88,11 @@ export class InvitesService {
     });
   }
 
-  async deleteInvite(id: string) {
+  async deleteInvite(id: string, userId: string) {
+    const invite = await this.prisma.invite.findUnique({ where: { id } });
+    if (invite.userId !== userId) {
+      throw new Error('You do not have permission to delete this invite');
+    }
     return this.prisma.invite.delete({ where: { id } });
   }
